requirements.txt
streamlit
streamlit-autorefresh
pandas
numpy
crcmod
pymodbus==3.6.8
paho-mqtt
openpyxl
